webapi/tct-xmlhttprequest-w3c-tests/xmlhttprequest/xhr2_XMLHttpRequest_response_byteLength_type.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
         Xu,Yuhan <yuhanx.xu@intel.com>
         Xu, Kang <kangx.xu@intel.com>

-->

<meta charset="utf-8">
<title>XMLHttpRequest Test: xhr2_XMLHttpRequest_response_byteLength_type</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/XMLHttpRequest/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function (t) {
  var client = new XMLHttpRequest();
  client.onload = t.step_func(function (event) {
    var data = this;
    assert_equals(data.status, 200, "The status value");
    assert_true(data.response instanceof ArrayBuffer, "ArrayBuffer data.response supported");
    assert_true("byteLength" in data.response, "data.response.byteLength supported");
    assert_equals(typeof data.response.byteLength, "number", "The type of XMLHttpRequest.response");
    t.done();
  });
  client.onerror = t.step_func(function () {
    assert_unreached("There is no trigger the onload event");
  });
  window.setTimeout(t.step_func(function () {
    assert_unreached("There is no trigger the onload event");
  }), 2000);
  client.open("GET", "support/test.xml");
  client.responseType = "arraybuffer";
  client.send("null");
}, "Check if the type of XMLHttpRequest.response is number");
</script>
